split real tikv test suites
