+++ .github/workflows/integration-test-with-real-tikv.yml
@@ -11,31 +11,18 @@ concurrency:
   cancel-in-progress: true
 
 jobs:
-  pessimistic-test:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - uses: actions/setup-go@v2
-        with:
-          go-version: 1.18
-      - name: Setup TiKV cluster
-        run: |
-          # Disable pipelined pessimistic lock temporarily until tikv#11649 is resolved
-          echo -e "[pessimistic-txn]\npipelined = false\n" > tikv.toml
-          echo -e "[raftdb]\nmax-open-files = 10240\n" >> tikv.toml
-          echo -e "[rocksdb]\nmax-open-files = 10240\n" >> tikv.toml
-          curl --proto '=https' --tlsv1.2 -sSf https://tiup-mirrors.pingcap.com/install.sh | sh
-          $HOME/.tiup/bin/tiup playground nightly --mode tikv-slim --kv 3 --pd 3 --without-monitor --kv.config tikv.toml &
-          curl --retry-connrefused --retry 5 --retry-delay 5 http://127.0.0.1:2379
-
-      - name: Run Tests
-        run: |
-          export log_level=error
-          make failpoint-enable
-          go test ./tests/pessimistictest -v -with-real-tikv -timeout 20m
-
   realtikv-test:
     runs-on: ubuntu-latest
+
+    strategy:
+      matrix:
+        suite:
+          - brietest
+          - pessimistictest
+          - sessiontest
+          - statisticstest
+          - txntest
+
     steps:
       - uses: actions/checkout@v2
       - uses: actions/setup-go@v2
@@ -54,4 +41,4 @@ jobs:
         run: |
           export log_level=error
           make failpoint-enable
-          go test ./tests/realtikvtest -v -with-real-tikv -timeout 30m
+          go test ./tests/realtikvtest/${{ matrix.suite }} -v -with-real-tikv -timeout 30m
