+++ tests/realtikvtest/txntest/isolation_test.go
@@ -12,12 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package realtikvtest
+package txntest
 
 import (
 	"testing"
 
 	"github.com/pingcap/tidb/testkit"
+	"github.com/pingcap/tidb/tests/realtikvtest"
 	"github.com/pingcap/tidb/util"
 	"github.com/stretchr/testify/require"
 )
@@ -27,7 +28,7 @@ These test cases come from the paper <A Critique of ANSI SQL Isolation Levels>.
 The sign 'P0', 'P1'.... can be found in the paper. These cases will run under snapshot isolation.
 */
 func TestP0DirtyWrite(t *testing.T) {
-	store, clean := createMockStoreAndSetup(t)
+	store, clean := realtikvtest.CreateMockStoreAndSetup(t)
 	defer clean()
 	session1 := testkit.NewTestKit(t, store)
 	session2 := testkit.NewTestKit(t, store)
@@ -86,7 +87,7 @@ func TestP0DirtyWrite(t *testing.T) {
 }
 
 func TestP1DirtyRead(t *testing.T) {
-	store, clean := createMockStoreAndSetup(t)
+	store, clean := realtikvtest.CreateMockStoreAndSetup(t)
 	defer clean()
 	session1 := testkit.NewTestKit(t, store)
 	session2 := testkit.NewTestKit(t, store)
@@ -136,7 +137,7 @@ func TestP1DirtyRead(t *testing.T) {
 }
 
 func TestP2NonRepeatableRead(t *testing.T) {
-	store, clean := createMockStoreAndSetup(t)
+	store, clean := realtikvtest.CreateMockStoreAndSetup(t)
 	defer clean()
 	session1 := testkit.NewTestKit(t, store)
 	session2 := testkit.NewTestKit(t, store)
@@ -207,7 +208,7 @@ func TestP2NonRepeatableRead(t *testing.T) {
 }
 
 func TestP3Phantom(t *testing.T) {
-	store, clean := createMockStoreAndSetup(t)
+	store, clean := realtikvtest.CreateMockStoreAndSetup(t)
 	defer clean()
 	session1 := testkit.NewTestKit(t, store)
 	session2 := testkit.NewTestKit(t, store)
@@ -275,7 +276,7 @@ func TestP3Phantom(t *testing.T) {
 }
 
 func TestP4LostUpdate(t *testing.T) {
-	store, clean := createMockStoreAndSetup(t)
+	store, clean := realtikvtest.CreateMockStoreAndSetup(t)
 	defer clean()
 	session1 := testkit.NewTestKit(t, store)
 	session2 := testkit.NewTestKit(t, store)
@@ -336,7 +337,7 @@ func TestP4LostUpdate(t *testing.T) {
 func TestP4CLostUpdate(t *testing.T) {}
 
 func TestA3Phantom(t *testing.T) {
-	store, clean := createMockStoreAndSetup(t)
+	store, clean := realtikvtest.CreateMockStoreAndSetup(t)
 	defer clean()
 	session1 := testkit.NewTestKit(t, store)
 	session2 := testkit.NewTestKit(t, store)
@@ -389,7 +390,7 @@ func TestA3Phantom(t *testing.T) {
 }
 
 func TestA5AReadSkew(t *testing.T) {
-	store, clean := createMockStoreAndSetup(t)
+	store, clean := realtikvtest.CreateMockStoreAndSetup(t)
 	defer clean()
 	session1 := testkit.NewTestKit(t, store)
 	session2 := testkit.NewTestKit(t, store)
@@ -454,7 +455,7 @@ func TestA5AReadSkew(t *testing.T) {
 }
 
 func TestA5BWriteSkew(t *testing.T) {
-	store, clean := createMockStoreAndSetup(t)
+	store, clean := realtikvtest.CreateMockStoreAndSetup(t)
 	defer clean()
 	session1 := testkit.NewTestKit(t, store)
 	session2 := testkit.NewTestKit(t, store)
@@ -551,7 +552,7 @@ These test cases come from the paper <Highly Available Transactions: Virtues and
 for tidb, we support read-after-write on cluster level.
 */
 func TestReadAfterWrite(t *testing.T) {
-	store, clean := createMockStoreAndSetup(t)
+	store, clean := realtikvtest.CreateMockStoreAndSetup(t)
 	defer clean()
 	session1 := testkit.NewTestKit(t, store)
 	session2 := testkit.NewTestKit(t, store)
@@ -604,7 +605,7 @@ func TestReadAfterWrite(t *testing.T) {
 This case will do harm in Innodb, even if in snapshot isolation, but harmless in tidb.
 */
 func TestPhantomReadInInnodb(t *testing.T) {
-	store, clean := createMockStoreAndSetup(t)
+	store, clean := realtikvtest.CreateMockStoreAndSetup(t)
 	defer clean()
 	session1 := testkit.NewTestKit(t, store)
 	session2 := testkit.NewTestKit(t, store)
